split up files and add some interfaces
