+++ modules/client/src/body.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useCallback, ReactElement } from 'react';
+import { useState, useCallback, ReactElement } from 'react';
 import { Visibility, Segment, Menu, Container, Grid } from 'semantic-ui-react';
 import { Link } from 'react-router-dom';
 import AuthButton from './components/auth/button';
