+++ modules/client/src/routes/events/form.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
-import React, {
+import {
   ReactElement,
   useCallback,
   useEffect,
