+++ modules/common/src/index.ts
@@ -7,6 +7,7 @@ export { default as asyncForEach } from './utils/asyncForEach';
 export { default as except } from './utils/except';
 export { default as getById } from './utils/getById';
 export { default as intersect } from './utils/intersect';
+export { default as clamper } from './utils/clamper';
 
 export { default as Roles } from './security/roles';
