+++ modules/common/src/utils/clamper.ts
@@ -1,0 +1,6 @@
+const { max, min } = Math;
+
+const clamper = (minimum: number, maximum: number) => (input: number): number =>
+  max(minimum, min(input, maximum));
+
+export default clamper;
